requirements.txt
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
filelock==3.15.4
fsspec==2024.6.1
gcloud==0.18.3
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.8
Jinja2==3.1.4
jwcrypto==1.5.6
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
openai-whisper==20231117
pillow==10.2.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.4
Pyrebase4==4.8.0
pysbd==0.3.4
python-jwt==4.1.0
regex==2024.7.24
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
semantic-version==2.10.0
setuptools==70.0.0
setuptools-rust==1.10.1
six==1.16.0
sympy==1.13.2
tiktoken==0.7.0
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
urllib3==1.26.19
